+++ SharpC2.sln
@@ -28,8 +28,8 @@ Global
 		{C45F4099-7EEA-476C-8989-0709018879D4}.Release|Any CPU.Build.0 = Release|Any CPU
 		{F0143C8C-EF61-4304-BBF2-DC52AFC3DE46}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F0143C8C-EF61-4304-BBF2-DC52AFC3DE46}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F0143C8C-EF61-4304-BBF2-DC52AFC3DE46}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F0143C8C-EF61-4304-BBF2-DC52AFC3DE46}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F0143C8C-EF61-4304-BBF2-DC52AFC3DE46}.Debug|Any CPU.ActiveCfg = Release|Any CPU
+		{F0143C8C-EF61-4304-BBF2-DC52AFC3DE46}.Debug|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 	EndGlobalSection
